feat(gitea): add hvcs support for gitea
Add Gitea ('gitea') to the list of supported hvcs'. Implementation is
basically a fork of the Github class without the Actions-part, as Gitea
is basically compatible to GitHub with some modifications when it comes
to the api.

Similar to the other hvcs implementations `GT_TOKEN` is used to provide
the Gitea api token, `hvcs_domain` sets the url to the Gitea instance,
`hvcs_api_domain` will be ignored as Giteas api url can be invoked from
`hvcs_domain`.